Move People forms/views into people dir for re-usability and create intake dir.
