+++ frontend/src/router.js
@@ -3,7 +3,9 @@ import Home from "./Home";
 import Animals, { AnimalDetail, NewAnimal, UpdateAnimal } from "./animals/Animals"
 import { AnimalSearch } from "./animals/AnimalTables"
 import Evac, { NewTeam, TeamList } from "./evac/Evac";
-import Hotline, { NewOwner, NewReporter, NewServiceRequest, OwnerDetail, ReporterDetail, ServiceRequestDetail, ServiceRequestList, UpdateOwner, UpdateReporter, UpdateServiceRequest } from "./hotline/Hotline";
+import Hotline, { NewServiceRequest, ServiceRequestDetail, ServiceRequestList, UpdateServiceRequest } from "./hotline/Hotline";
+import Intake, { IntakeSummary } from "./intake/Intake";
+import { NewOwner, NewReporter, OwnerDetail, ReporterDetail,  UpdateOwner, UpdateReporter } from "./people/People";
 import { Login } from "./accounts/Accounts"
 import { EvacTeamTable } from "./evac/EvacTables";
 import { EvacTeamForm } from "./evac/EvacForms";
@@ -29,6 +31,10 @@ const routes = {
   "/hotline/servicerequest/list": () => <ServiceRequestList />,
   "/hotline/servicerequest/new": () => <NewServiceRequest />,
   "/hotline/servicerequest/:id": ({id}) => <ServiceRequestDetail id={id}/>,
+  "/intake": () => <Intake />,
+  "/intake/owner/new": () => <NewOwner />,
+  "/intake/reporter/new": () => <NewReporter />,
+  "/intake/summary": () => <IntakeSummary />,
   "/login": () => <Login />,
 };
